Changed some types to unsigned as it makes no sense to have negative numbers there

+++ x_view_core/include/x_view_core/features/graph.h
@@ -17,10 +17,10 @@ namespace x_view {
 struct VertexProperty {
 
   /// \brief Index of the vertex in the graph.
-  /// \note This index should not be used to acces the vertex in the graph,
+  /// \note This index should not be used to access the vertex in the graph,
   /// as even if the vertex order is maintained, due to vertex merging
   /// operations it might happen that vertices are deleted from the graph.
-  int index;
+  uint64_t index;
 
   /// \brief Semantic label associated to this graph vertex. This label
   /// corresponds to the same specified in the dataset description.
@@ -53,7 +53,7 @@ struct EdgeProperty {
   /// const VertexProperty from_v_p = graph[from_v_d];
   /// \endcode
 
-  int from;
+  uint64_t from;
   /// \brief Index of the second vertex defining this edge.
   /// \note This index should not be used to access the associated vertex, as
   /// it might differ with the storage index of the graph. To access the
@@ -62,7 +62,7 @@ struct EdgeProperty {
   /// const VertexDescriptor to_v_d = boost::target(edge, graph);
   /// const VertexProperty to_v_p = graph[to_v_d];
   /// \endcode
-  int to;
+  uint64_t to;
 
   /// \brief Integer indicating how many times this edge has been observed.
   uint64_t num_times_seen;
